requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.4
Django==4.2.17
django-encrypted-searchable-fields==0.2.0
django-environ==0.12.0
django-filter==24.3
django-model-utils==5.0.0
django-oauth-toolkit==3.0.1
django-phone-field==1.8.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
environ==1.0
google-auth==2.40.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.4
rpds-py==0.22.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
